Input: cs40l26 - Restore PLL properly if open-loop
set_pll_loop() checks if the driver needs to lock the REFCLK
input PLL (closed-loop)  or to put it in free-running
(open-loop) mode.

The function was checking to see if the input parameter
(pll_loop) was equal to 0 or 1. However, the input
was only masked when set_pll_loop() is called from
vibe_state_update() at the end of an ASP event; it was
not shifted to be properly compared directly to 1.

Use FIELD_GET to mask and shift bit 11 of the REFCLK_INPUT
value so that it is not rejected in set_pll_loop() if the
PLL needs to be put in open-loop mode.

+++ include/linux/mfd/cs40l26.h
@@ -1264,7 +1264,7 @@ struct cs40l26_pll_sysclk_config {
 
 /* exported function prototypes */
 int cs40l26_svc_le_estimate(struct cs40l26_private *cs40l26, unsigned int *le);
-int cs40l26_set_pll_loop(struct cs40l26_private *cs40l26, unsigned int pll_loop);
+int cs40l26_set_pll_loop(struct cs40l26_private *cs40l26, u8 pll_loop);
 int cs40l26_asp_start(struct cs40l26_private *cs40l26);
 int cs40l26_num_ram_waves(struct cs40l26_private *cs40l26);
 int cs40l26_num_owt_waves(struct cs40l26_private *cs40l26);
